% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_req_auth}
\alias{client_req_auth}
\title{Authenticate the request with the OAuth client}
\usage{
client_req_auth(req)
}
\arguments{
\item{req}{The request object to authenticate.}
}
\value{
The authenticated request object.
}
\description{
Authenticate the request with the OAuth client
}
\keyword{internal}
