% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_api_url}
\alias{get_api_url}
\title{Get API URL from environment variable}
\usage{
get_api_url(path = NULL)
}
\arguments{
\item{path}{A character string representing the API URL.}
}
\value{
A character string representing the API URL.
}
\description{
This function retrieves the API URL from the environment variable named "CONSIBIO_API_HOST", if it is set.
If the environment variable is not set, it set and returns a default URL.
Optionally, it can append a path to the URL.
}
\examples{
\dontrun{
get_api_url("/users/me")
}

}
