% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get_device.R
\name{get_device}
\alias{get_device}
\title{Get device in project}
\usage{
get_device(device_id = NULL)
}
\arguments{
\item{device_id}{The device id.}
}
\value{
A matrix with the device information.
}
\description{
Get the project device in Consibio APIs.
}
\details{
See details in https://api.v2.consibio.com/api-docs/#/default/get_projects__project_id__devices__device_id_
}
\examples{
\dontrun{
get_device(device_id = "{device_id}")
}
}
