% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get_devices.R
\name{get_devices}
\alias{get_devices}
\title{Get devices in project}
\usage{
get_devices()
}
\value{
A matrix with the devices information.
}
\description{
Get the project devices in Consibio APIs.
}
\details{
See details in https://api.v2.consibio.com/api-docs/#/default/get_projects__project_id__devices
}
\examples{
\dontrun{
get_devices()
}
}
