% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get_element.R
\name{get_element}
\alias{get_element}
\title{Get element information}
\usage{
get_element(element_id = NULL)
}
\arguments{
\item{element_id}{The element id.}
}
\value{
A matrix with the element information.
}
\description{
Get the element information from Consibio APIs.
}
\details{
See details in https://api.v2.consibio.com/api-docs/#/default/get_projects__project_id__elements__entity_id_
}
\examples{
\dontrun{
get_element(element_id = "{element_id}")
}
}
