% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get_elements.R
\name{get_elements}
\alias{get_elements}
\title{Get elements accessible to the user}
\usage{
get_elements()
}
\value{
A matrix with the project information.
}
\description{
Get the projects elements to the user from Consibio APIs.
}
\details{
See details in https://api.v2.consibio.com/api-docs/#/default/get_projects__project_id__elements
}
\examples{
\dontrun{
get_elements()
}
}
