% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get_project.R
\name{get_project}
\alias{get_project}
\title{Get project information}
\usage{
get_project(project_id = NULL)
}
\arguments{
\item{project_id}{The project id (optional).}
}
\value{
A matrix with the project information.
}
\description{
Get the project information from Consibio APIs.
}
\details{
See details in https://api.v2.consibio.com/api-docs/#/default/get_project
}
\examples{
\dontrun{
get_project(project_id = "{project_id}")
}
}
