% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_project_id}
\alias{get_project_id}
\title{Get the Consibio project ID}
\usage{
get_project_id()
}
\value{
The Consibio project ID
}
\description{
This function retrieves the Consibio project ID from the options.
If the project ID is not set or is an empty string, an error is thrown.
}
\examples{
# Set the Consibio project ID
# Alternatively, you can use the set_project_id('PROJECT_ID') function
options(consibio.project_id = "PROJECT_ID")

# Get the Consibio project ID
get_project_id()
}
