% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get projects accessible to the user}
\usage{
get_projects()
}
\value{
A matrix with the project information.
}
\description{
Get the projects accessible to the user from Consibio APIs.
}
\details{
See details in https://api.v2.consibio.com/api-docs/#/default/get_projects
}
\examples{
\dontrun{
get_projects()
}
}
