% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get_users_me.R
\name{get_users_me}
\alias{get_users_me}
\title{Get the user information}
\usage{
get_users_me()
}
\value{
A matrix with the user information.
}
\description{
Get the user information from Consibio APIs.
}
\details{
See details in https://api.v2.consibio.com/api-docs/#/default/get_users_me
}
\examples{
\dontrun{
get_users_me()
}
}
