% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_entity_id}
\alias{is_valid_entity_id}
\title{Is valid entity ID}
\usage{
is_valid_entity_id(entity_id = NULL, type = NULL)
}
\arguments{
\item{entity_id}{The entity ID}

\item{type}{The type of the entity ID. Default is 'project' (will just change the halt error message)}
}
\value{
TRUE if the entity ID is valid, FALSE otherwise
}
\description{
Is valid entity ID
}
\examples{
\dontrun{
is_valid_entity_id("123456", "project")
}
}
