% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_username}
\alias{is_valid_username}
\title{Check if a username is valid.}
\usage{
is_valid_username(username = NULL)
}
\arguments{
\item{username}{The username to be checked.}
}
\value{
TRUE if the username is valid, otherwise an error is thrown.
}
\description{
By using regular expressions, this function checks if a username is a valid email address.
}
\examples{
is_valid_username("john.doe@example.com")
\dontrun{
is_valid_username("invalid_username")
is_valid_username("")
is_valid_username(NULL)
}

}
