% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{login}
\alias{login}
\title{Login to Consibio API}
\usage{
login(username = NULL)
}
\arguments{
\item{username}{A character string representing the username.}
}
\value{
A character string representing the logged-in username.
}
\description{
Use username to login to the Consibio API. Password will be requested interactively, now and once the token expires.
It sets the default username for authenticated requests.
}
\examples{
\dontrun{
login("bob@helloworld.com")
}

}
