% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_project_id}
\alias{set_project_id}
\title{Set the Consibio project ID option}
\usage{
set_project_id(project_id)
}
\arguments{
\item{project_id}{The project ID to set.}
}
\value{
The project ID that was set.
}
\description{
Set the Consibio project ID option
}
\examples{
set_project_id("PROJECT_ID")
}
